backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
python-multipart
mysqlclient
sqlalchemy
python-dotenv
alembic
pyodbc